Finishing touches for network pool provider
The current implementation of the nutanix prism element plugin doesn't
support creating or editing network pools. As such, there's nothing much
to translate from the embedded that didn't come along with the seed
translation. This commit just tidies up a few of the calls and puts
comments in to make it clear we're intentionally not supporting these
calls.

#66109 #66110 #66111 #66112