Remove IPAMProvider and use seed for NetworkPoolType
The network pool provider was largely a stub so that we could provide
a NetworkPoolType for the plugin. By registering the NetworkPoolProvider
we ended up showing menu items around morpheus that were unintended and
not supported. Instead, let's just seed the NetworkPoolType and remove
the registration of the NetworkPoolProvider for now. In the future, we
can flesh out support for the various methods of an IPAM provider and
actually make use of it.